EndDayScreen won't let you move on if you dont pay debt, or have enough money

+++ SciFiGame/Assets/Scripts/MainManagerScript.cs
@@ -10,8 +10,8 @@ public class MainManager : MonoBehaviour
     public static MainManager Instance;
 
     //Player stats
-    public int debt;
-    public int money;
+    public float debt;
+    public float money;
     public int rArmHealth;
     public int lArmHealth;
     public int rLegHealth;
